backend/src/tests/services/UserService.test.ts
import https from 'https';
import fetch from "node-fetch";
import { UserDto } from 'common/src/models/UserDto';
import { Config } from '../../Config';
import { EntitiesDataSource } from '../../data/EntitiesDataSource';
import { UserEntity } from '../../data/UserEntity';

jest.setTimeout(Config.jestTimeoutSeconds * 1000);

describe("UsersService", () => {
    let agent = new https.Agent({ rejectUnauthorized: false });
    let entityGuid = "faf76b3d-ed66-4182-a7c2-7ea6562785fe";
    let token: string | undefined;
    let eds: EntitiesDataSource;

    beforeAll(async () => {
        eds = new EntitiesDataSource();
        await eds.initialize();

        const body = JSON.stringify({
            emailAddress: "administrator@localhost",
            password: "Welcome123"
        });

        const response = await fetch("https://localhost:4433/api/v0/auth/login", {
            agent: agent,
            method: "POST",
            body: body,
            headers: { "Content-Type": "application/json" }
        });

        if (!response.ok)
            throw new Error(`Response: ${response.status} - ${response.statusText}`);

        const obj = await response.json();
        if (!obj["data"])
            throw new Error("Return from login did not provide a 'data' with the token!");

        token = obj["data"] as string;
    }, Config.jestTimeoutSeconds * 1000);
    afterAll(async () => {
        try { await eds.userRepository().delete({ guid: entityGuid }); }
        finally { await eds.destroy(); }
    }, Config.jestTimeoutSeconds * 1000);

    test("POST /api/v0/user - save new should return 200", async () => {
        if (!token)
            throw new Error("No token - did beforeAll() fail?");

        const entity = new UserEntity();
        entity.guid = entityGuid;
        entity.displayName = "Delete Me";
        entity.emailAddress = "deleteme@localhost";
        entity.smsPhone = "555-555-5555";

        const response = await fetch("https://localhost:4433/api/v0/user", {
            agent: agent,
            method: "POST",
            body: JSON.stringify(entity),
            headers: {
                "Content-Type": "application/json",
                "Authorization": "Bearer " + token
            }
        });

        if (!response.ok)
            throw new Error(`Response: ${response.status} - ${response.statusText}`);

        expect(response.ok).toBeTruthy();
        expect(response.status).toBe(200);

        let reloaded = await eds.userRepository().findOneByOrFail({ guid: entityGuid });
        expect(entity).toEqual(reloaded);
    }, Config.jestTimeoutSeconds * 1000);
    test("POST /api/v0/user overwrite should return 200", async () => {
        if (!token)
            throw new Error("No token - did beforeAll() fail?");

        const entity = new UserEntity();
        entity.guid = entityGuid;
        entity.displayName = "Delete Me Dupe";
        entity.emailAddress = "deleteme@localhost";
        entity.smsPhone = "555-555-5555";

        const response = await fetch("https://localhost:4433/api/v0/user", {
            agent: agent,
            method: "POST",
            body: JSON.stringify(entity),
            headers: {
                "Content-Type": "application/json",
                "Authorization": "Bearer " + token
            }
        });

        if (!response.ok)
            throw new Error(`Response: ${response.status} - ${response.statusText}`);

        expect(response.ok).toBeTruthy();
        expect(response.status).toBe(200);

        let reloaded = await eds.userRepository().findOneByOrFail({ guid: entityGuid });
        expect(entity).toEqual(reloaded);
    }, Config.jestTimeoutSeconds * 1000);
    test("GET /api/v0/users should return user list", async () => {
        if (!token)
            throw new Error("No token - did beforeAll() fail?");

        const response = await fetch("https://localhost:4433/api/v0/users", {
            agent: agent,
            method: "GET",
            headers: {
                "Content-Type": "application/json",
                "Authorization": "Bearer " + token
            }
        });

        if (!response.ok)
            throw new Error(`Response: ${response.status} - ${response.statusText}`);

        const obj = await response.json();
        if (!obj["data"])
            throw new Error("No data returned!");

        const data = obj["data"] as UserDto[];

        expect(data.length > 0).toBeTruthy();
        expect(data[0].guid).toBeTruthy();
        expect(data[0].displayName).toBeTruthy();
        expect(data[0].emailAddress).toBeTruthy();
        expect(data[0].smsPhone).toBeTruthy();
    }, Config.jestTimeoutSeconds * 1000);
    test("GET /api/v0/user/:guid should return user and 200", async () => {
        if (!token)
            throw new Error("No token - did beforeAll() fail?");

        const response = await fetch("https://localhost:4433/api/v0/user/" + entityGuid, {
            agent: agent,
            method: "GET",
            headers: {
                "Content-Type": "application/json",
                "Authorization": "Bearer " + token
            }
        });

        if (!response.ok)
            throw new Error(`Response: ${response.status} - ${response.statusText}`);

        const obj = await response.json();
        if (!obj["data"])
            throw new Error("No data returned!");

        const data = obj["data"] as UserDto;

        expect(data.guid).toEqual(entityGuid);
    }, Config.jestTimeoutSeconds * 1000);
    test("DELETE /api/v0/user/:guid should delete user and return 200", async () => {
        if (!token)
            throw new Error("No token - did beforeAll() fail?");

        const response = await fetch("https://localhost:4433/api/v0/user/" + entityGuid, {
            agent: agent,
            method: "DELETE",
            headers: {
                "Content-Type": "application/json",
                "Authorization": "Bearer " + token
            }
        });

        if (!response.ok)
            throw new Error(`Response: ${response.status} - ${response.statusText}`);

        expect(response.ok).toBeTruthy();
        expect(response.status).toBe(200);

        let entity = await eds.userRepository().findBy({ guid: entityGuid });
        expect(entity.length).toEqual(0);
    }, Config.jestTimeoutSeconds * 1000);
});